遠征ログの読み込み部分を外出し
